requirements_production.txt
# Core dependencies
Flask==2.3.3
Flask-CORS==4.0.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Production server
gunicorn==21.2.0

# Security
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
Flask-SeaSurf==1.1.1

# Monitoring & Logging
prometheus-client==0.17.1
structlog==23.2.0

# Caching (optional)
Flask-Caching==2.1.0
redis==5.0.1

# Health checks
healthcheck==1.3.3

# Environment management
python-dotenv==1.0.0

# Development tools (optional for production)
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0
